requirements.txt
# MCP Server dependencies
mcp>=1.0.0
fastmcp>=0.1.0

# HTTP client for QlikSense API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# JSON handling and data processing
pydantic>=2.0.0

# Async support
asyncio-mqtt>=0.16.0

# Logging
structlog>=23.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0